[resolves #1889] Cannot obtain direct consumer ...
...with spring context

The meaning of `noStart` flag on `SpringCamelContext` was slightly
changed with CAMEL-11261, it blocks any `SpringCamelContext` startup
even the one invoked explicitly via `start` method.

This changes the creation of SpringCamelContext so that the `noStart` is
set and maintained while Spring application context is refreshed, but
reset to true after.